refactor endpoint v2 expr json helpers
